Fix a caching bug in base taggers
The AsyncTagger and BasicTagger implementations were incorrectly
handling edit scenarios.  Instead of extending the CachedSpan value to
include the requests from the new ITextSnapshot with the previous
CachedSpan value it just used the new requset in it's entirety.  This
lead to a scenario where the tagger thought it was only responsible for
a single line while the editor thought it was responsible for many other
lines.  Hence when the tagger later "cleared" tags it only cleared one
line.

This is best explained through an example
- Buffer has 10 lines
- AsyncTagger returned a tag for every line in the buffer
- User pasted a new line into the editor at line 5
- Editor asked AsyncTagger for tags for line 5 only
- AsyncTagger responded correctly but then set it's CachedSpan to line 5
- IAsyncTaggerSource later raised "Changed" and AsyncTagger then raised
Changed with only line 5

At this point the AsyncTagger thinks it's cleared all tags.  But in
reality it only cleared line 5 and the editor happily maintained the
tags on lines 1-4 and 6-11