Work around double load problem
It turns out that Visual Studio will double load a MEF component if 2
VSIX projects deploy it as a MEF dll even if the item is strongly named.

Both test projects in this solution both reference + deploy EditorUtils
as a MEF component.  Hence on F5 of either EditorUtils was being loaded
twice in MEF which caused all of the singleton components to be exported
twice.  This broke pretty much everything in EditorUtils and any code
referencing it.

After some investigation it appears the safe way to do this is to deploy
the shared library as it's own VSIX.  Then it is loaded once.  For
development purposes the libraries continue to reference EditorUtils
they just don't list it as a MEF component.