Added VSIX to NuGet Deployment
This change adds the generated VSIX file to the EditorUtils NuGet
deployment.  It also updates the Deploy.ps1 script to check for certain
mistakes in the packaging (mainly version number consistency).

Ideally though this should be extended to automatically update the
source.extension.vsixmanifest to properly deploy EditorUtilsVsix.vsix as
well.