Efficient GetTags code
The GetTags code no longer tries to process Tag values for the
overarching span provided to GetTags.  Instead it focuses on the actual
SnapshotSpan values that were provided in the collection.  This will
increase efficiency in places where we have collapsed regions or
projection buffers (most common cases for this)