Moved to simpler host model
There is now a clean separation between the creation of the host and the
actual usage of the host.  The mixed mode was causing a large
performance issue in testing because the host was being forced to
re-query the same CompositionContainer over and over for the same
values.  There is no need for this, once set the values should be the
same.